Update subjects page
Now subjects are managed using api on client

+++ quizer/api/views.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from main.models import Subject, Test
-from main import mongo
+from main import mongo, utils
 from .serializers import SubjectSerializer, TestSerializer
 from .permissions import IsLecturer
 
@@ -26,32 +26,47 @@ class SubjectView(APIView):
             'subjects': serializer.data
         })
 
-    def post(self, request):
-        subject = request.data.get('subject')
-        serializer = SubjectSerializer(data=subject)
-        if serializer.is_valid(raise_exception=True):
-            new_subject = serializer.save()
-        return Response({
-            'success': "Предмет '%s' успешно добавлен." % new_subject.name
-        })
+    def post(self, request, pk):
+        request_dict = dict(request.POST)
+        if len(request_dict) == 1:  # DELETE, only csrftoken passed
+            subject = get_object_or_404(Subject.objects.all(), pk=pk)
+            subject_name = subject.name
+            tests = Test.objects.filter(subject__id=subject.id)
+            tests_count = tests.count()
 
-    def put(self, request, pk):
-        updated_subject = get_object_or_404(Subject.objects.all(), pk=pk)
-        data = request.data.get('subject')
-        serializer = SubjectSerializer(instance=updated_subject, data=data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            updated_subject = serializer.save()
-        return Response({
-            'success': "Предмет '%s' был успешно отредактирован." % updated_subject.name
-        })
+            deleted_questions_count = 0
+            storage = mongo.QuestionsStorage.connect(db=mongo.get_conn())
+            for test in tests:
+                deleted_questions_count += storage.delete_many(test_id=test.id)
+            subject.delete()
 
-    def delete(self, _, pk):
-        subject = get_object_or_404(Subject.objects.all(), pk=pk)
-        subject_name = subject.name
-        subject.delete()
-        return Response({
-            'success': "Предмет '%s' был успешно удален." % subject_name
-        }, status=204)
+            message = "Учебный предмет '%s', %d тестов к нему, а также все " + \
+                 "вопросы к тестам в количестве %d были успешно удалены."
+            return Response({
+                'success': message % (subject_name, tests_count, deleted_questions_count)
+            })
+        elif pk == 'new':  # POST
+            serializer = SubjectSerializer(data=request.data)
+            print(serializer)
+            if serializer.is_valid(raise_exception=True):
+                new_subject = serializer.save()
+            return Response({
+                'success': "Предмет '%s' успешно добавлен." % new_subject.name
+            })
+        elif pk == 'load':  # POST
+            print(request.POST)
+            message = utils.add_subject_with_tests(request)
+            return Response({
+                'success': message
+            })
+        else:  # PUT
+            updated_subject = get_object_or_404(Subject.objects.all(), pk=pk)
+            serializer = SubjectSerializer(instance=updated_subject, data=request.data, partial=True)
+            if serializer.is_valid(raise_exception=True):
+                updated_subject = serializer.save()
+            return Response({
+                'success': "Предмет '%s' был успешно отредактирован." % updated_subject.name
+            })
 
 
 class TestView(APIView):
